rename NewGrid to Grid
